+++ examples/Light/index.js
@@ -27,7 +27,7 @@ const {
   RESULT_SUCCESS,
   RESULT_FAILURE,
   ThingAccessClient,
-  getConfig,
+  Config,
 } = require('linkedge-thing-access-sdk');
 
 // Max retry interval in seconds for registerAndOnline.
@@ -88,7 +88,7 @@ class Connector {
    * Connects to Link IoT Edge and publishes properties to it.
    */
   connect() {
-    registerAndOnlineWithBackOffRetry(this._client, 1)
+    return registerAndOnlineWithBackOffRetry(this._client, 1)
       .then(() => {
         return new Promise(() => {
           // Publish properties to Link IoT Edge.
@@ -109,7 +109,7 @@ class Connector {
    * Disconnects from Link IoT Edge and stops publishing properties to it.
    */
   disconnect() {
-    this._client.cleanup()
+    return this._client.cleanup()
       .catch(err => {
         console.log(err);
       });
@@ -170,15 +170,15 @@ class Connector {
 }
 
 // Get the config which is auto-generated when devices are bound to this driver.
-getConfig()
+Config.get()
   .then((config) => {
     // Get the device information from config, which contains product key, device
     // name, etc. of the device.
-    const things = config.getThings();
-    things.forEach((thing) => {
+    const thingInfos = config.getThingInfos();
+    thingInfos.forEach((thingInfo) => {
       const light = new Light();
       // The Thing format is just right for connector config, pass it directly.
-      const connector = new Connector(thing, light);
+      const connector = new Connector(thingInfo, light);
       connector.connect();
     });
   });